vet_ai/pyproject.toml
[project]
name = "vet-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.0",
    "google-genai>=1.24.0",
    "logfire>=3.22.1",
    "streamlit>=1.46.1",
]
